finished backend

+++ backend/controllers/goalController.js
@@ -1,13 +1,13 @@
-const asyncHandler = require("express-async-handler"); // instead of try-catch block ( look  npm  asyncHandler)
+const asyncHandler = require("express-async-handler"); // instead of try-catch block ( look  asyncHandler)
 
 const Goal = require("../models/goalsModel");
-const { request } = require("express");
+const User = require("../models/userModel");
 
 // Get goals
 // route GET /api/goals
 // private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({user: req.user.id});
 
   res.status(200).json(goals);
 });
@@ -22,6 +22,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -36,6 +37,16 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+  const user = await User.findById(req.user.id)
+
+  if(!user){
+    res.status(401)
+    throw new Error('user not found')
+  }
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -53,6 +64,17 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const delGoal = await Goal.findByIdAndRemove({ _id: req.params.id });
